GUI v3 problema date solucionado

+++ Stats/Core/Session.cs
@@ -4,7 +4,7 @@ namespace Stats
 {
 	public class Session
 	{
-		public int start{ get; set;}
+		public long start{ get; set;}
 		public int duration{ get; set;}
 		public int distance{ get; set;}
 		public int weight{ get; set;}
